Improve handling of decimals, implement dot button
The dot button enables the user to type in decimal numbers for
calculation.
Further I replaced the wrongly chosen parseInt() method with the more
appropriate parseFloat() method to enable the application to properly
calculate decimals.
Lastly I limited the presentation of decimal to 5 places to keep the
number inside the display element.